refactor(music): extract the search and playlist page components out of their screens
